fluid clean: remove parallel and parallel_helper api
