+++ python/paddle/fluid/dygraph/__init__.py
@@ -21,10 +21,6 @@ from .layers import *
 from . import tracer
 from .tracer import *
 
-from . import parallel
-from .parallel import *
-
-
 from . import learning_rate_scheduler
 from .learning_rate_scheduler import *
 
@@ -33,5 +29,4 @@ from .math_op_patch import monkey_patch_math_varbase
 __all__ = []
 __all__ += layers.__all__
 __all__ += base.__all__
-__all__ += parallel.__all__
 __all__ += learning_rate_scheduler.__all__
